Add clear and fix concat with empty heaps
Fix a problem when using concat with empty heaps due to a bug in the
circular doubly linked list implementation which left a reference to the
previous list sentinel on a node when combining lists.

Add a new clear method that empties the heap much like the existing
methods on Array and Hash.

Simplify implementation where possible (e.g. not resetting the root list
when unnecessary) and port across an optimisation from the reference
text to create an array of the appropriate size when consolidating
trees. This relies on the property that the biggest degree will always
be less than or equal to the binary log of n ("lg n").